added heatmap with userdata and sentiment barchart

+++ BotDetection/App_Data/SingleTweet.cs
@@ -11,7 +11,7 @@ namespace BotDetection
         public int retweets { get; set; }
         public int likes { get; set; }
         public ArrayList Hashtags { get; set; }
-        public int sentiment { get; set; }
+        public double sentiment { get; set; }
         public int tweetID { get; set; }
 
         public SingleTweet()
@@ -19,7 +19,7 @@ namespace BotDetection
             Content = "";
             Hashtags = new ArrayList();
             PostTime = new DateTime();
-            sentiment = 0;
+            sentiment = 0.0;
             retweets = 0;
             likes = 0;
             tweetID = 0;
